design-audit: extract shared pendingCardSx, move PendingBadge to common/
FeedItem and ObservationGridCard each reimplemented the same "dim +
disable interaction" treatment for pending optimistic cards with
diverging mechanics (opacity on Card vs CardActionArea, onClick guard
vs preventDefault). PendingBadge's own doc comment already called out
the pairing as a repeated contract. Extract pendingCardSx() into
PendingBadge.tsx, and move it to common/ since ObservationGridCard
(itself in common/) already depended on it.
